Disable scrollbar
I am getting used to read the percent in the modeline. I never liked it
because the height of the scrollbar handle varies depending on the
amount of characters shown, not depending on the number of lines shown.

I already do most of movements through the buffer through the keyboard anyway.