Set default font depending on available system fonts
Added a list of fonts, the first font that is available in the system
will be used and the rest ignored.

At work (Windows environment) I do not have DejaVu Sans Mono, so I
removed it from Custom settings and configured it directly.

I think this approach is much better than having a different setup for
ech different system, since I do not have to list the systems explicitly.