Initialize package on new systems
Initialize package system when user's ELPA repository is empty and
there are no packages installed.

It is important because Emacs only initializes package module if there are
installed packages. Otherwise it must be initialized manually.

In addition, this will update ELPA archive if it is empty and install use-package.

To improve:
 * What happens if there is no internet?
 * Should ELPA archive be refreshed before installing use-package?