Use #' when functions are passed as arguments
"The compiler throws a warning whenever it notices you've used an
undefined function, say (not-defined "oops"), but it can't do the same
for something like (mapcar 'not-defined some-list) because it doesn't
know that symbol is the name of a function. The sharp quote is a way of
conveying that information to the compiler, so if it runs into
(mapcar #'not-defined some-list), it can throw a warning accordingly.

"So it is always good practice to sharp quote every symbol that is the
name of a function, whether it's going into a mapcar, an apply, a
funcall, or anything else. Adhering to this actually unearthed a small
bug in one of my packages."

Source:
http://endlessparentheses.com/get-in-the-habit-of-using-sharp-quote.html