Set a better value for scroll conservatively
The previous value, 101, disabled this feature and 100% of the time
scrolled by one line. The disadvantage is that C-s or other jump
commands would leave the cursor at the top or bottom of the window.

With the new value, 2, most of the time when scrolling it moves one
line. In addition, when jumping to another position, it centers the
cursor in the window.