Ediff - Do not use a new frame
Ediff usually creates a new Emacs frame where ediff menu is shown. This
frame gets in the middle and it's hard to get back there if it's hidden
by the main Emacs frame.

Disabling the frame and using a window instead solves the problem, even
taking into account the extra space used.