Use always Dejavu Sans Mono as font
By using always the same font, the configuration is now working even
when running Emacs as a daemon.

If the font is not properly selected, just install it from
http://dejavu-fonts.org/.

In the previous code a check was made to ensure that the font was
available to Emacs before setting it. When emacs is run as a daemon this
check always fails because the GUI is not initialized until there is a
graphical frame (this is by design); causing the new frames to have the
default font instead of the customization.

The font size may be parameterized depending on the operating system if
needed. To be checked on OSX.