Add function to remove all properties in a buffer
When copying and pasting emacs text between buffers in different modes,
sometimes the pasted text will show colors and fonts that do not apply
in the new buffer.

This function removes all the properties from the text in the buffer and
then Emacs can color it properly according to the current mode.