Remove functions to open a file/dir as root
These functions were used directly from the minibuffer or from ido when finding
a file or a directory.

It is not needed anymore because is integrated into counsel. When finding a
file or dir using counsel, press M-o instead of enter or tab. A menu will be
shown with additional options, one of these options is open as root.

Unfortunately Tramp for Administrator access does not work in Windows.